requirements.txt
certifi==2017.11.5
Django==1.11.5
django-braces==1.12.0
django-cors-middleware==1.3.1
django-extensions==1.7.8
django-model-utils==3.0.0
django-oauth-toolkit==1.0.0
djangorestframework==3.6.2
fake-factory==0.5.3
oauthlib==2.0.1
pep8==1.7.0
pytz==2018.3
six==1.11.0
virtualenv==12.1.1
psycopg2-binary==2.7.4
gunicorn==19.7.1